+++ discovery-provider/src/api/v1/users.py
@@ -77,11 +77,6 @@ from src.queries.get_followees_for_user import get_followees_for_user
 from src.queries.get_followers_for_user import get_followers_for_user
 from src.queries.get_related_artists import get_related_artists
 from src.queries.get_repost_feed_for_user import get_repost_feed_for_user
-from src.queries.get_save_tracks import (
-    GetTrackLibraryArgs,
-    LibraryFilterType,
-    get_track_library,
-)
 from src.queries.get_saves import get_saves
 from src.queries.get_subscribers import (
     get_subscribers_for_user,
@@ -94,6 +89,11 @@ from src.queries.get_support_for_user import (
 from src.queries.get_top_genre_users import get_top_genre_users
 from src.queries.get_top_user_track_tags import get_top_user_track_tags
 from src.queries.get_top_users import get_top_users
+from src.queries.get_track_library import (
+    GetTrackLibraryArgs,
+    LibraryFilterType,
+    get_track_library,
+)
 from src.queries.get_tracks import GetTrackArgs, get_tracks
 from src.queries.get_unclaimed_id import get_unclaimed_id
 from src.queries.get_user_listen_counts_monthly import get_user_listen_counts_monthly
